fix: - Исправлено по замечаниям

+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -5,13 +5,11 @@ import ru.yandex.practicum.filmorate.exceptions.StorageException;
 import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Review;
 import ru.yandex.practicum.filmorate.storage.dao.*;
-import ru.yandex.practicum.filmorate.storage.impl.DbFeedDaoImpl;
 
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -54,9 +52,6 @@ public class ReviewService {
     }
 
     public Review save(Review review) {
-        if (review.getUserId() == null || review.getFilmId() == null) {
-            throw new IllegalStateException("Не заполнены поля filmId или userId");
-        }
 
         if ((userDao.findUserById(review.getUserId()).isPresent())
                 && (filmDao.findFilmById(review.getFilmId()).isPresent())) {
@@ -92,10 +87,14 @@ public class ReviewService {
     }
 
     public boolean delete(long id) {
-        Optional<Review> review = reviewDao.findReviewById(id);
-        feedDao.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
-                review.get().getUserId(), "REVIEW", "REMOVE", review.get().getReviewId()));
-        return reviewDao.delete(id);
+        if (reviewDao.findReviewById(id).isPresent()) {
+            Review review = reviewDao.findReviewById(id).get();
+            feedDao.saveFeed(new Feed(1, Instant.now().toEpochMilli(),
+                    review.getUserId(), "REVIEW", "REMOVE", review.getReviewId()));
+            return reviewDao.delete(id);
+        } else {
+            return false;
+        }
     }
 
     public void addLike(long reviewId, long userId) {
